Large refactor: use Constructor to instantiate, then delegate.
Still not sure what the relationship should be between the constructor
and the runner. Also, I think 'step' might be oversimplifying things in
the runner, since there is two parts to the update for PSO.  One, to
calculate the best position, best fitness etc - and another to calculate
the new velocity + position.